requirements.txt
-f https://download.pytorch.org/whl/nightly/cpu/torch_nightly.html
torch_nightly
fastai
starlette
uvicorn
python-multipart
aiofiles
google-cloud-storage
databases[sqlite]